pyproject.toml
[tool.poetry]
name = "autonomous-trading-system"
version = "0.1.0"
description = "Multi-LLM powered autonomous trading system"
authors = ["Trading Team <team@example.com>"]
readme = "README.md"
packages = [
    {include = "apps"},
    {include = "libs"},
]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
httpx = "^0.25.0"
pandas = "^2.1.0"
numpy = "^1.25.0"
scipy = "^1.11.0"
# TA-Lib = "0.4.28"  # Commented out due to build issues on Windows
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"
redis = "^5.0.0"
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
# opentelemetry-instrumentation = "^0.41b0"  # Optional monitoring dependency
psutil = "^5.9.0"
pyyaml = "^6.0.0"
structlog = "^23.2.0"
cryptography = "^41.0.0"
pyjwt = "^2.8.0"
websockets = "^12.0"
aiohttp = "^3.9.0"

# New dependencies for optimization
toml = "^0.10.2"
watchdog = "^3.0.0"
tenacity = "^8.2.3"
backoff = "^2.2.1"
slowapi = "^0.1.9"
python-dotenv = "^1.0.0"  # For native deployment .env file support
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
ruff = "^0.1.0"
mypy = "^1.7.0"
black = "^23.11.0"
pre-commit = "^3.5.0"
bandit = "^1.7.5"
safety = "^2.3.0"
hypothesis = "^6.88.0"

# Performance testing
pytest-benchmark = "^4.0.0"
pytest-profiling = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "S",  # bandit
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "S101",  # use of assert detected
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=apps --cov=libs --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "slow: marks tests as slow running",
    "asyncio: marks tests as async tests",
    "performance: marks tests as performance tests",
]

[tool.coverage.run]
source = ["apps", "libs"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.benchmark]
min_rounds = 5
warmup = true